Move files to implement setting page
